+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### dev-master
+
+* #3: escape(null) returns empty string not type-error
+
 ### 1.0.0 (07.07.2019)
 
 * `HTMLBuilder` class